backend/requirements.txt
fastapi
uvicorn
pydantic
spacy
scikit-learn
numpy
pandas
joblib
pdfplumber
python-docx
nltk
